% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarySegmentationCPDetection.R
\name{binarySegmentationCPDetection}
\alias{binarySegmentationCPDetection}
\title{Binary Segmentation for Change Point Detection}
\usage{
binarySegmentationCPDetection(
  fullData,
  method,
  lower,
  upper,
  alpha,
  estimVarLS = FALSE,
  var1BaseEstim = 0.5,
  var2BaseEstim = NA,
  maxOptimIters = 1,
  trimAmt = 4,
  verifyCPs = TRUE,
  silent = F
)
}
\arguments{
\item{fullData}{Data.frame with two columns. The first has date and the
second has values}

\item{method}{String indicating method to use for change point detection.
Options are 'MLE', 'Vostrikova' , and 'WLS'.}

\item{lower}{Vector of numerics indicating the lower bound. Variance cannot
be negative and the third term must also be above 0.}

\item{upper}{Vector of numerics indicating the upper bound.}

\item{alpha}{Numeric indicating the significance of interest.}

\item{estimVarLS}{(Optional) Boolean to indicate if we should use LS to
estimate var2. Default is FALSE.}

\item{var1BaseEstim}{(Optional) Numeric indicating base var1 estimate. Default
is 0.5.}

\item{var2BaseEstim}{(Optional) Numeric indicating base var1 estimate. Default
is NA. May not be used if estimVarLS.}

\item{maxOptimIters}{(Optional) Numeric indicating number of times to re-try
optimization. Can sometimes add if get randomly odd result. Default is 1.}

\item{verifyCPs}{(Optional) Boolean indicating if change points should be
rechecked at the end. Default is TRUE.}

\item{silent}{(Optional) Boolean indicating if output should be supressed.
Default is FALSE}
}
\value{
Data.frame with columns 'stDate' (start date of region), 'enDate'
(end date of region), 'WLS_b1' (beta1 estimate using WLS), 'WLS_v1' (var1
estimate using WLS), 'WLS_v2' (var2 estimate using WLS), 'WLS_b2' (beta2
estimate using WLS), 'MLE_b1' (beta1 estimate using MLE), 'MLE_v1' (var1
estimate using MLE), 'MLE_v2' (var2 estimate using MLE), and 'MLE_b2'
(beta2 estimate using MLE).
}
\description{
This function uses binary segmentation to detect change points.
}
\examples{
data <- generateRCA1Data(pars=c(0.5,1,0.25,-0.5), k=250, burnin=1000,
                        iterations=500)
binarySegmentationCPDetection(data,'Vostrikova',
                              lower=c(-Inf,0,10^-8,-Inf),
                              upper=c(Inf,Inf,Inf,Inf), alpha=0.05)
}
