% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTrim.R
\name{computeTrim}
\alias{computeTrim}
\title{Compute Trim}
\usage{
computeTrim(requireDrops, type, N = NA)
}
\arguments{
\item{requireDrops}{Integer indicating the number of require dropped. Ensure
this is at least enough to estimate the parameters. If NA, use log(N)
for trim (min of 4)}

\item{type}{String of Start or End. This indicates if this is the start or
ending trim}

\item{N}{(Optional) Integer of the length of data. Needed if type='End'}
}
\value{
Numeric indicating the index at which to start/stop examining in the
data
}
\description{
This is just a function to control how much is trimmed
}
\examples{
computeTrim(10,'Start')
computeTrim(10,'End',100)
}
