% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeWLS.R
\name{computeWLSVals}
\alias{computeWLSVals}
\title{Compute Weighted Least Square Values}
\usage{
computeWLSVals(y, N = length(y), trim = log(N))
}
\arguments{
\item{y}{Vector of numerics for the data}

\item{N}{(Optional) Numeric indicating length of the data. Included to
increase speed. Default is length(y).}

\item{trim}{(Optional) Numeric indicating trim amount. Default is log(N).}
}
\value{
Vector of two values: value and location of max
}
\description{
This uses weighted least squares to estimate paramters for RCA(1) with changes.
This is for homoskedastic data. This work is replication of that done
in Horvath, Trapani 2022.
}
\examples{
data <- generateRCA1Data(pars=c(0.5,1,0.25,-0.5), k=500, burnin=1000,
                        iterations=1000)
computeWLSVals(data)
}
