% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRCA1Data.R
\name{.checkNonStationary}
\alias{.checkNonStationary}
\title{Check Non stationarity}
\usage{
.checkNonStationary(pars, stationarySims, stationaryCutoff)
}
\arguments{
\item{pars}{Vector of parameters for the normal errors,
c(beta1,sigma1^2,sigma2^2,beta2). Only beta1 and sigma1^2 are used.}

\item{stationarySims}{Integer for the number of simulations used to check
stationarity.}

\item{stationaryCutoff}{Numeric indicating the cutoff to determine
stationarity. True cutoff would be 0, but due to finite samples or wanting
to permit some nonstationarity or only want very stationary process, can
make influence user to choose different number.}
}
\value{
Boolean indicating if the process is nonstationary (T) or stationary (F)
}
\description{
This (internal) function check if parameters (for normally distributed errors)
are in the stationary or nonstationary regime.
}
\details{
Upcoming: Check when non-normal. Simple mathematically, just need to get
around to it.
}
\examples{
# This is an internal function and will not be viewable. See use in
#     generateData.
}
