% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCutoff.R
\name{.computeCutoff}
\alias{.computeCutoff}
\title{Compute Cutoff}
\usage{
.computeCutoff(alpha, N, type, nStart = NA, nEnd = NA)
}
\arguments{
\item{alpha}{Numeric indicating the significance of interest.}

\item{N}{Numeric indicating the length of data}

\item{type}{String indicating type of cutoff. Options are 'General' or
'Vostrikova'.}

\item{nStart}{(Optional) Numeric indicating starting value for trim. Default
is NA.}

\item{nEnd}{(Optional) Numeric indicating ending value for trim. Default
is NA.}
}
\value{
Numeric indicating the cutoff value
}
\description{
An (internal) function to compute the cutoff value for the statistic of interest.
}
\examples{
# This is an internal function and will not be viewable. See use in
#     compute3Methods.
}
