% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRCA1Data.R
\name{.dataGen}
\alias{.dataGen}
\title{Data Generation}
\usage{
.dataGen(
  pars,
  burnin,
  iterations,
  errorType,
  par1HetereoLocation,
  par1HetereoMult,
  par2HetereoLocation,
  par2HetereoMult,
  start = NA
)
}
\arguments{
\item{pars}{Vector of parameters for the errors. The format is always similar,
c(beta1, p1, p2); however, p1 and p2 depend on the errorType.
Errors are always centered around 0
\itemize{
\item Normal: beta1, sigma1^2, sigma2^2
\item Bernoulli: beta1, prob1, prob2
\item Exponential: beta1, rate1, rate2
}}

\item{burnin}{Integer indicating the number of burnin iteration to use}

\item{iterations}{Integer indicating how many observations the final data
should be.}

\item{errorType}{String indicating the type of error. Options are 'Normal',
'Bernoulli', and 'Exponential'.}

\item{par1HetereoLocation}{(Optional) This can be used to allow
hetereoskedasticity in the par1 (2nd term in pars) error. This determines
location for it.}

\item{par1HetereoMult}{(Optional) This can be used to allow
hetereoskedasticity in the par1 (2nd term in pars) error. This determines
the amount to multiply par1 by for it.}

\item{par2HetereoLocation}{(Optional) This can be used to allow
hetereoskedasticity in the par2 (3rd term in pars) error. This determines
location for it.}

\item{par2HetereoMult}{(Optional) This can be used to allow
hetereoskedasticity in the par2 (3rd term in pars) error. This determines
the amount to multiply par2 by for it.}

\item{start}{(Optional) Numeric, if not NA, will start from this value}
}
\value{
Vector with the sequence of RCA data of length iterations according
to the params
}
\description{
This (internal) function generates data according to an RCA process, using
the given parameters and information.
}
\examples{
# This is an internal function and will not be viewable. See use in
#     generateData.
}
