% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarySegmentationCPDetection.R
\name{.detectChangePoints}
\alias{.detectChangePoints}
\title{Detect Change Points}
\usage{
.detectChangePoints(
  data,
  method,
  lower,
  upper,
  alpha,
  addAmt = 0,
  estimVarLS = FALSE,
  var1BaseEstim = 0.5,
  var2BaseEstim = 0.5,
  maxOptimIters = 1,
  trimAmt = 4,
  silent = F
)
}
\arguments{
\item{data}{Vector of numerics with values, pre-ordered}

\item{method}{String indicating method to use for change point detection.
Options are 'MLE', 'Vostrikova' , and 'WLS'.}

\item{lower}{Vector of numerics indicating the lower bound. Variance cannot
be negative and the third term must also be above 0.}

\item{upper}{Vector of numerics indicating the upper bound.}

\item{alpha}{Numeric indicating the significance of interest.}

\item{addAmt}{(Optional) Numeric to add to CPs so can use this function
recursively.}

\item{estimVarLS}{(Optional) Boolean to indicate if we should use LS to
estimate var2. Default is FALSE.}

\item{var1BaseEstim}{(Optional) Numeric indicating base var1 estimate. Default
is 0.5.}

\item{var2BaseEstim}{(Optional) Numeric indicating base var1 estimate. Default
is 0.5. If NA will compute based on robust statistics. May not be used if
estimVarLS.}

\item{maxOptimIters}{(Optional) Numeric indicating number of times to re-try
optimization. Can sometimes add if get randomly odd result. Default is 1.}

\item{silent}{(Optional) Boolean indicating if output should be supressed.
Default is FALSE}
}
\value{
Vector of integers indicating the change points
}
\description{
Detect Change Points
}
\examples{
# This is an internal function. See use in binarySegmentationCPDetection.
}
