% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateMaxLikelihood.R
\name{.estimateMaxLikelihood}
\alias{.estimateMaxLikelihood}
\title{Estimate Maximum Likelihood}
\usage{
.estimateMaxLikelihood(
  y,
  k,
  estimates,
  lower,
  upper,
  maxOptimIters = 2,
  returnWithVal = FALSE
)
}
\arguments{
\item{y}{Vector of numerics for the data}

\item{k}{Integer indicating the point at when the process should change from
using beta1 to using beta2}

\item{estimates}{Vector of 4 numerics indicating initial est}

\item{lower}{Vector of numerics indicating the lower bound. Variance cannot
be negative and the third term must also be above 0.}

\item{upper}{Vector of numerics indicating the upper bound.}

\item{maxOptimIters}{(Optional) Numeric indicating number of times to re-try
optimization. Can sometimes add if get randomly odd result. Default is 2.}

\item{returnWithVal}{(Optional) Boolean indicating if the estimated parameters
should be returned with the estimated values.}
}
\value{
Depends on returnWithVal
\itemize{
\item If TRUE: Vector of the estimated parameters and the estimated
value using the parameters.
\item If FALSE: Vector of the estimated parameters
}
}
\description{
An (internal) function that optimizes the functions for parameters based on
the data with the changepoint.
}
\examples{
# This is an internal function and will not be viewable. See use in
#     computeTN.
}
