% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateMaxLikelihood.R
\name{.estimateMaxLikelihood_array}
\alias{.estimateMaxLikelihood_array}
\title{Estimate Maximum Likelihood - Array}
\usage{
.estimateMaxLikelihood_array(
  y,
  k,
  estim1,
  estim2,
  estim3,
  estim4,
  lower,
  upper,
  maxOptimIters = 2,
  returnWithVal = FALSE
)
}
\arguments{
\item{y}{List of vector of numerics for the data}

\item{k}{Integer indicating the point at when the process should change from
using estim1 to using estim4}

\item{estim1}{Vector of numerics for estimates of beta1}

\item{estim2}{Vector of numerics for estimates of sigma1^2}

\item{estim3}{Vector of numerics for estimates of sigma2^2}

\item{estim4}{Vector of numerics for estimates of beta2}

\item{lower}{List of vector of numerics indicating the lower bound. Variance
cannot be negative and the third term must also be above 0.}

\item{upper}{List of vector of numerics indicating the upper bound.}

\item{maxOptimIters}{(Optional) Numeric indicating number of times to re-try
optimization. Can sometimes add if get randomly odd result. Default is 2.}

\item{returnWithVal}{(Optional) Boolean indicating if the estimated parameters
should be returned with the estimated values.}
}
\value{
Depends on returnWithVal
\itemize{
\item If TRUE: Vector of the estimated parameters and the estimated
value using the parameters.
\item If FALSE: Vector of the estimated parameters
}
}
\description{
An (internal) function that optimizes the functions for parameters based on
the data with the changepoint. Organizes to use as array for speed.
}
\examples{
# This is an internal function and will not be viewable. See use in
#     computeTN.
}
