% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCutoff.R
\name{.findX}
\alias{.findX}
\title{Find X}
\usage{
.findX(x, alpha, eAlpha)
}
\arguments{
\item{x}{Numeric x to consider}

\item{alpha}{Numeric indicating the significance of interest.}

\item{eAlpha}{Stardardized N^2 to handle values discussed 2.8}
}
\value{
Numeric indicating x value
}
\description{
This (internal) function is used to compute x as needed in Vostrikova
improvement of cutoff.
}
\examples{
# This is an internal function and will not be viewable. See use in
#     .VostrikovaCutoff.
}
