% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCutoff.R
\name{.GeneralCutoff}
\alias{.GeneralCutoff}
\title{General Cutoff}
\usage{
.GeneralCutoff(alpha, N)
}
\arguments{
\item{alpha}{Numeric indicating the significance of interest.}

\item{N}{Numeric indicating the length of data}
}
\value{
Numeric indicating the cutoff value
}
\description{
Cutoff given in Theorem 2.1.
}
\examples{
# This is an internal function and will not be viewable. See use in
#     .computeCutoff.
}
