% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateStationaryRegion.R
\name{.generateEdge}
\alias{.generateEdge}
\title{Generate Edge}
\usage{
.generateEdge(data, silent = FALSE)
}
\arguments{
\item{data}{Data.frame indicating stationarity at each beta-epsilon1 value.
The columns are beta (num), var1 (num), val (num), stable (bool).}

\item{silent}{(Optional) Boolean indicating if output should be supressed (T).
Default is FALSE.}
}
\value{
Data,frame of data with column boundary (Bool) being bound indicating
if the beta-var1 is a boundary position
}
\description{
This (internal) function finds the edges of the stationary region.
}
\examples{
# This is an internal function. See use in simulateStationaryRegion.
}
