% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateStationaryRegion.R
\name{.generateStationaryRegion_memory}
\alias{.generateStationaryRegion_memory}
\title{Generate Stationary Region - Memory}
\usage{
.generateStationaryRegion_memory(bSeq, epsSeq, nSims)
}
\arguments{
\item{bSeq}{Vector indicating sequence of beta values to be examined}

\item{epsSeq}{Vector indicating sequence of epsilon (1) values to be examined}

\item{nSims}{Integer indicating the number of simulation at each point}
}
\value{
Data.frame indicating stationarity at each beta-epsilon1 value. The
columns are beta (num), var1 (num), val (num), stable (bool).
}
\description{
This (internal) function computes stationarity of region using Monte Carlo
simulation. This strictly uses matrices rather than loops. May have
memory issues.
}
\examples{
# This is an internal function. See use in simulateStationaryRegion.
}
