% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateStationaryRegion.R
\name{.generateStationaryRegion_theo}
\alias{.generateStationaryRegion_theo}
\title{Generate Stationary Region - Theoretical}
\usage{
.generateStationaryRegion_theo(bSeq, epsSeq, nSims, silent = F)
}
\arguments{
\item{bSeq}{Vector indicating sequence of beta values to be examined}

\item{epsSeq}{Vector indicating sequence of epsilon (1) values to be examined}

\item{nSims}{Integer indicating the number of simulation at each point}

\item{silent}{(Optional) Boolean indicating if output should be supressed (T).
Default is FALSE.}
}
\value{
Data.frame indicating stationarity at each beta-epsilon1 value. The
columns are beta (num), var1 (num), val (num), stable (bool).
}
\description{
This (internal) function computes stationarity of region using Monte Carlo
simulation. This using quantiles rather than random computing MC
simulations.
}
\examples{
# This is an internal function. See use in simulateStationaryRegion.
}
