% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLikelihoods.R
\name{.L}
\alias{.L}
\title{Log Liklihood}
\usage{
.L(u, k, y, N = length(y))
}
\arguments{
\item{u}{Vector of numerics for parameters, c(beta1, sigma1^2, sigma2^2, beta2)}

\item{k}{Integer indicating the value to stop using beta1 (i.e. the change point)}

\item{y}{Vector of numerics for the data}

\item{N}{(Optional) Integer indicating the length of y. Value added to
increase speed rather than calling length(y) each call.}
}
\value{
Numeric of log liklihood with change point k
}
\description{
This (internal) function computes the log liklihood, defined as 2.1 in the paper
}
\examples{
# This is an internal function.
}
