% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLikelihoods.R
\name{.L_array}
\alias{.L_array}
\title{(Array) Log Liklihood}
\usage{
.L_array(u1, u2, u3, u4, k, y, N = length(y))
}
\arguments{
\item{u1}{Numeric for parameter 1, beta1}

\item{u2}{Numeric for parameter 2, sigma1^2}

\item{u3}{Numeric for parameter 3, sigma2^2}

\item{u4}{Numeric for parameter 4, beta2}

\item{k}{Integer indicating the value to stop using beta1 (i.e. the change point)}

\item{y}{Vector of numerics for the data}

\item{N}{(Optional) Integer indicating the length of y. Value added to
increase speed rather than calling length(y) each call.}
}
\value{
Numeric of log liklihood with change point k
}
\description{
This (internal) function computes the log liklihood, defined as 2.1 in the
paper. This is built to better fit for array calls.
}
\examples{
# This is an internal function.
}
