% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeWLS.R
\name{.LSVarestimate}
\alias{.LSVarestimate}
\title{Least Squares Variance1 Estimate}
\usage{
.LSVarestimate(y, beta1, beta2 = NA, k = NA)
}
\arguments{
\item{y}{List of vector of data}

\item{beta1}{Numeric for beta1 estimate}

\item{beta2}{(Optional) Numeric for beta2 estimate}

\item{k}{(Optional) Integer location of change point when beta2 is present}
}
\value{
Numeric indicating the var1 estimate
}
\description{
This (internal) function computes the least squares estimate of variance1
}
\examples{
# This is an internal function. See for example binarySegmentationCPDetection.
}
