% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarySegmentationCPDetection.R
\name{.verifyChangePoints}
\alias{.verifyChangePoints}
\title{Verify Change Points}
\usage{
.verifyChangePoints(
  CPsVals,
  data,
  method,
  lower,
  upper,
  alpha,
  estimVarLS,
  var1BaseEstim,
  var2BaseEstim,
  maxOptimIters,
  trimAmt,
  silent = F
)
}
\arguments{
\item{CPsVals}{Vector of integers indicating the detected change points}

\item{data}{Vector of numerics indicating data values}

\item{method}{String indicating method to use for change point detection.
Options are 'MLE', 'Vostrikova' , and 'WLS'.}

\item{lower}{Vector of numerics indicating the lower bound. Variance cannot
be negative and the third term must also be above 0.}

\item{upper}{Vector of numerics indicating the upper bound.}

\item{alpha}{Numeric indicating the significance of interest.}

\item{estimVarLS}{Boolean to indicate if we should use LS to estimate var2.}

\item{var1BaseEstim}{Numeric indicating base var1 estimate.}

\item{var2BaseEstim}{Numeric indicating base var1 estimate.}

\item{maxOptimIters}{Numeric indicating number of times to re-try
optimization. Can sometimes add if get randomly odd result.}

\item{silent}{(Optional) Boolean indicating if output should be supressed.
Default is FALSE}
}
\value{
Vector of integers indicating the change points
}
\description{
Verify Change Points
}
\examples{
# This is an internal function. See use in binarySegmentationCPDetection.
}
