% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeWLS.R
\name{.WLSestimate}
\alias{.WLSestimate}
\title{Weighted Least Squares Estimate}
\usage{
.WLSestimate(y, k, lower, upper)
}
\arguments{
\item{y}{Vector of numerics for the data}

\item{k}{Integer(s) location of change point. Can do this as a vector}

\item{lower}{Vector of numerics indicating the lower bound. Variance cannot
be negative and the third term must also be above 0.}

\item{upper}{Vector of numerics indicating the upper bound.}
}
\value{
WLS estimates for beta1, var1, var2, beta2 at potentially different k
}
\description{
Weighted Least Squares Estimate
}
\examples{
# Internal function. See for use example binarySegmentationCPDetection.
}
