% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeWLS.R
\name{.WLSMeansEstimate}
\alias{.WLSMeansEstimate}
\title{Weighted Least Squares Mean Estimates}
\usage{
.WLSMeansEstimate(y, k)
}
\arguments{
\item{y}{Vector of numerics for the data}

\item{k}{Integer location of change point}
}
\value{
Vector of two numerics indicating beta1 and beta2 estimates
}
\description{
Compute weighted least squares mean estimates.
}
\examples{
# Internal function, see use in .WLSestimate.
}
