% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeWLS.R
\name{.WLSVarsEstim}
\alias{.WLSVarsEstim}
\title{Weighted Least Squares Variance Estimates}
\usage{
.WLSVarsEstim(y, changepoint, means)
}
\arguments{
\item{y}{List of vector of numerics of data}

\item{changepoint}{Integer location of change point}

\item{means}{Vector of two numerics indicating means. Likely from
.WLSMeansEstimate.}
}
\value{
Vector of two numerics indicating var1 and var2 estimates
}
\description{
Compute weighted least squares variance estimates. See for example books like
Random Coefficient Autoregressive Models: An Introduction (Des F. Nicholls,
Barry G. Quinn) or A New Iterative Procedure for Estimation of RCA
Parameters Based on Estimating Functions (Norlli Anida Abdullah et al.)
}
\examples{
# Internal function, see use in .WLSestimate.
}
