% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing.R
\docType{data}
\name{housing}
\alias{housing}
\title{US Housing Market Index}
\format{
A data frame with 39,535 rows and 3 columns:
\describe{
\item{city}{US cityv}
\item{date}{Date}
\item{index}{Housing market index}
}
}
\source{
\url{https://doi.org/10.1002/jae.2471}
}
\usage{
housing
}
\description{
Data from showing the housing index in several markets.
}
\keyword{datasets}
