% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderStackedPlot.R
\name{renderStackedPlot}
\alias{renderStackedPlot}
\title{Render Stacked Plot}
\usage{
renderStackedPlot(trueData, parCPData, title, subtitle, varPlots = TRUE)
}
\arguments{
\item{trueData}{Data.frame of data, the first column is date and the second
column is the value}

\item{parCPData}{Data.frame of change points. First column is start point,
second column is end point, third is beta value, fourth is variance1, and
fifth is variance2.}

\item{title}{String to be title of the plot}

\item{subtitle}{String to be subtitle of the plot}

\item{varPlots}{(Optional) Boolean to indicate if variance plots should also
be built. Default is TRUE.}
}
\value{
GGplot indicating data wiht change points, beta values, and
potentially variance values
}
\description{
This function renders a stacked plot based on data and change points
}
\examples{
\dontrun{
data <- UKcovid[UKcovid$nation=='Wales',c(2,3)]
data <- data[order(data$date),]
}

data <- generateRCA1Data(pars=c(0.5,1,0.25,-0.5),
                         k=100, burnin=1000,
                         iterations=200)

result_Vost <- detectAllChangePoints(fullData=data,
                       method='Vostrikova',
                       lower=c(-Inf, 0, 10^-8, -Inf),
                       upper=c(Inf, Inf, Inf, Inf),
                       alpha=0.05, nStart=NA, nEnd=NA)
data_plot <- stackedPlot(trueData = data,
                       parCPData = result_Vost[,c(1:2,7:10)],
                       title = NULL, subtitle = NULL,
                       dataName = nation, varPlots = FALSE)
}
