% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateStationaryRegion.R
\name{simulateStationaryRegion}
\alias{simulateStationaryRegion}
\title{Simulate Stationary Region}
\usage{
simulateStationaryRegion(
  nSims = 10000,
  addEdge = TRUE,
  generationTrick = NA,
  silent = F
)
}
\arguments{
\item{nSims}{(Optional) Integer indicating the number of simulations to use
at each point}

\item{addEdge}{(Optional) Boolean indicating if the edge should be drawn}

\item{generationTrick}{(Optional) String that can indicate tricks to improve
running time. Default is NA.
\itemize{
\item NA: No tricks, just run. This will take the most time
\item theoretical: Rather than generating normals, use quantiles to
get theoretical values
\item reuse: Rather than generating unique normals at each point, use
one set for all
\item memory: Uses complete matrix multiplication rather than loops.
Very fast, but likely will have memory issues on most machines
for large nSims
}}

\item{silent}{(Optional) Boolean indicating if output should be suppressed (T).
Default is FALSE.}
}
\value{
List with data.frame of date, (opt) data.frame of edge information,
and ggplot2 plot.
}
\description{
This simulates and draws the RCA(1) stationary region
}
\examples{
dat <- simulateStationaryRegion()
}
