% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationTable_NC.R
\name{simulationTable_NC}
\alias{simulationTable_NC}
\title{Simulation Table with Null Case}
\usage{
simulationTable_NC(
  betas,
  varProbRates,
  nSims,
  iterations,
  burnin,
  lowerEstim,
  upperEstim,
  alpha,
  errorTypes,
  CPLoc = 0.5,
  seed = NA,
  silent = F,
  trimAmt = 4
)
}
\arguments{
\item{betas}{Vector of numerics indicating betas to examine (remember this is
the null case so only one beta used each iteration)}

\item{varProbRates}{Vector of 2 values indicating the variance or probabilitise
used in parameters, c(u2, u3).}

\item{nSims}{Integer indicating the number of simulations for each trial}

\item{iterations}{Integer indicating how many observations the final data
should be.}

\item{burnin}{Integer indicating the number of burnin iteration to use}

\item{lowerEstim}{Vector of numerics indicating the lower bound for the 4
parameters. Variance cannot be negative and the third term must also be
above 0.}

\item{upperEstim}{Vector of numerics indicating the upper bound for the 4
parameters.}

\item{alpha}{Numeric indicating the significance of interest.}

\item{errorTypes}{Vector indicating the errorTypes to investigate. Options
are 'Normal', 'Bernoulli', and 'Exponential'.}

\item{CPLoc}{Numeric indicating change point location, between 0 and 1.}

\item{seed}{(Optional) Integer to indicate seed to set at each interation.
Default of NA sets no seed.}

\item{silent}{(Optional) Boolean indicating if progress should be output.
Default is FALSE}

\item{trimAmt}{(Optional) Number indicating the amount to trim. Default is 4.}
}
\value{
Data.frame of null rejection frequencies
}
\description{
This function is used to generate a table of null rejection frequencies
}
\examples{
data_er <- simulationTable_NC(betas = c(0.5),
                              varProbRates = c(0.5, 0.5),
                              nSims=5, iterations = c(50),
                              burnin = 1000, lowerEst=c(-Inf,0,10^-8,-Inf),
                              upperEst=c(Inf,Inf,Inf,Inf), alpha = 0.05,
                              errorTypes = c('Normal'),
                              CPLoc=0.5)

\dontrun{
data_er <- simulationTable_NC(betas = c(0.5,0.75,1,1.05),
                              varProbRates = c(0.5, 0.5),
                              nSims=500, iterations = c(100, 200, 400, 800),
                              burnin = 1000, lowerEst=c(-Inf,0,10^-8,-Inf),
                              upperEst=c(Inf,Inf,Inf,Inf), alpha = 0.05,
                              errorTypes = c('Normal','Bernoulli','Exponential'),
                              CPLoc=0.5, seed=1234)
}
}
