% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKcovid.R
\docType{data}
\name{UKcovid}
\alias{UKcovid}
\title{UK Covid Data}
\format{
A data frame with 2614 rows and 3 columns:
\describe{
\item{nation}{Nation in the UK. The options are England, Northern Ireland,
Scotland, and Wales.}
\item{date}{Date}
\item{cases}{Number of reported hospitalisations}
}
}
\source{
\url{https://coronavirus.data.gov.uk/details/healthcare}
}
\usage{
UKcovid
}
\description{
Data showing hospitalisations in four UK nations.
}
\keyword{datasets}
